Using sinon also in nextcloud test file Return null if no event can be found to be compatible with the google controller
